Create Roblox KillBrick
This is my first published code for Roblox Lua. I will describe here what it does and how it works.
Create a Part, then create a local variable called "Part" which we will use as a name for the Part (script.Parent)
By using an event called "Touched" we can check if the Player touches the part.
We need to connect a function that will do something when the event is called (When the player is touching the part).
We need to put up a parameter with any name that will store which other part touched it.
So IF "Humanoid" (Player) touches the Part THEN we set the humanoid (Player) health to 0, the player dies.